# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: RSS Reader
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths
basePath: /
#
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /feed:
    x-swagger-router-controller: feed
    post:
      tags:
        - feed
      summary: Add a new feed
      description: ''
      operationId: addFeed
      produces:
        - application/json
      parameters:
      - name: url
        in: query
        description: Feed URL
        required: true
        type: string
      responses:
        '200':
          description: Success
        '400':
          description: Bad Request
        '405':
          description: Invalid input
  '/feed/{feedId}':
    x-swagger-router-controller: feed
    get:
      tags:
        - feed
      summary: Find feed by ID
      description: Returns a single feed
      operationId: getFeedById
      produces:
        - application/json
        #- text/plain
      parameters:
        - name: feedId
          in: path
          description: ID of feed to return
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Feed'
        '400':
          description: Invalid ID supplied
        '404':
          description: Feed not found
  '/feed/{feedId}/items':
    x-swagger-router-controller: feed
    get:
      tags:
        - feed
      summary: Get feed items
      description: Returns array of feed items
      operationId: getFeedItems
      produces:
        - application/json
        #- text/plain
      parameters:
        - name: feedId
          in: path
          description: ID of feed to return
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Item'
        '400':
          description: Invalid ID supplied
        '404':
          description: Feed not found
  # This serves swagger file (without sections that are swagger extensions - x-*)
  /swagger:
    x-swagger-pipe: swagger_raw
definitions:
  Feed:
    type: object
    properties:
      id:
        type: integer
        format: int64
      status:
        type: string
        enum:
          - OK
          - STOPPED
      url:
        type: string
      title:
        type: string
      added:
        type: string
        format: date-time
  Item:
    type: object
    properties:
      id:
        type: integer
        format: int64
      guid:
        type: string
      feedId:
        type: integer
        format: int64
      title:
        type: string
      url:
        type: string
      description:
        type: string
      enclosureUrl:
        type: string
      enclosureType:
        type: string
      date:
        type: string
        format: date-time
      added:
        type: string
        format: date-time